# doc/api.yaml
openapi: 3.0.0
paths:
  "/user":
    get:
      operationId: UsersController_findAll
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - Users
    post:
      operationId: UsersController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateUserDto"
      responses:
        '201':
          description: ''
      tags:
        - Users
  "/user/{id}":
    get:
      operationId: UsersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          example: 68280d15-5093-4e0e-80fd-147e796e252c
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - Users
    put:
      operationId: UsersController_update
      parameters:
        - name: id
          required: true
          in: path
          example: 68280d15-5093-4e0e-80fd-147e796e252c
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdatePasswordDto"
      responses:
        '200':
          description: ''
      tags:
        - Users
    delete:
      operationId: UsersController_delete
      parameters:
        - name: id
          required: true
          in: path
          example: 68280d15-5093-4e0e-80fd-147e796e252c
          schema:
            type: string
      responses:
        '204':
          description: ''
      tags:
        - Users
  "/artist":
    get:
      operationId: ArtistsController_findAll
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - Artists
    post:
      operationId: ArtistsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateArtistDto"
      responses:
        '201':
          description: ''
      tags:
        - Artists
  "/artist/{id}":
    get:
      operationId: ArtistsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          example: 68280d15-5093-4e0e-80fd-147e796e252c
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - Artists
    put:
      operationId: ArtistsController_update
      parameters:
        - name: id
          required: true
          in: path
          example: 68280d15-5093-4e0e-80fd-147e796e252c
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateArtistDto"
      responses:
        '200':
          description: ''
      tags:
        - Artists
    delete:
      operationId: ArtistsController_delete
      parameters:
        - name: id
          required: true
          in: path
          example: 68280d15-5093-4e0e-80fd-147e796e252c
          schema:
            type: string
      responses:
        '204':
          description: ''
      tags:
        - Artists
  "/album":
    get:
      operationId: AlbumsController_findAll
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - Albums
    post:
      operationId: AlbumsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateAlbumDto"
      responses:
        '201':
          description: ''
      tags:
        - Albums
  "/album/{id}":
    get:
      operationId: AlbumsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          example: 68280d15-5093-4e0e-80fd-147e796e252c
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - Albums
    put:
      operationId: AlbumsController_update
      parameters:
        - name: id
          required: true
          in: path
          example: 68280d15-5093-4e0e-80fd-147e796e252c
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateAlbumDto"
      responses:
        '200':
          description: ''
      tags:
        - Albums
    delete:
      operationId: AlbumsController_delete
      parameters:
        - name: id
          required: true
          in: path
          example: 68280d15-5093-4e0e-80fd-147e796e252c
          schema:
            type: string
      responses:
        '204':
          description: ''
      tags:
        - Albums
  "/track":
    get:
      operationId: TracksController_findAll
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - Tracks
    post:
      operationId: TracksController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateTrackDto"
      responses:
        '201':
          description: ''
      tags:
        - Tracks
  "/track/{id}":
    get:
      operationId: TracksController_findOne
      parameters:
        - name: id
          required: true
          in: path
          example: 68280d15-5093-4e0e-80fd-147e796e252c
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - Tracks
    put:
      operationId: TracksController_update
      parameters:
        - name: id
          required: true
          in: path
          example: 68280d15-5093-4e0e-80fd-147e796e252c
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateTrackDto"
      responses:
        '200':
          description: ''
      tags:
        - Tracks
    delete:
      operationId: TracksController_delete
      parameters:
        - name: id
          required: true
          in: path
          example: 68280d15-5093-4e0e-80fd-147e796e252c
          schema:
            type: string
      responses:
        '204':
          description: ''
      tags:
        - Tracks
  "/favs":
    get:
      operationId: FavoritesController_findAll
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - Favorites
  "/favs/track/{id}":
    post:
      operationId: FavoritesController_addFavoriteTrack
      parameters:
        - name: id
          required: true
          in: path
          example: 68280d15-5093-4e0e-80fd-147e796e252c
          schema:
            type: string
      responses:
        '201':
          description: ''
      tags:
        - Favorites
    delete:
      operationId: FavoritesController_deleteTrackFromFavorites
      parameters:
        - name: id
          required: true
          in: path
          example: 68280d15-5093-4e0e-80fd-147e796e252c
          schema:
            type: string
      responses:
        '204':
          description: ''
      tags:
        - Favorites
  "/favs/album/{id}":
    post:
      operationId: FavoritesController_addFavoriteAlbum
      parameters:
        - name: id
          required: true
          in: path
          example: 68280d15-5093-4e0e-80fd-147e796e252c
          schema:
            type: string
      responses:
        '201':
          description: ''
      tags:
        - Favorites
    delete:
      operationId: FavoritesController_deleteAlbumFromFavorites
      parameters:
        - name: id
          required: true
          in: path
          example: 68280d15-5093-4e0e-80fd-147e796e252c
          schema:
            type: string
      responses:
        '204':
          description: ''
      tags:
        - Favorites
  "/favs/artist/{id}":
    post:
      operationId: FavoritesController_addFavoriteArtist
      parameters:
        - name: id
          required: true
          in: path
          example: 68280d15-5093-4e0e-80fd-147e796e252c
          schema:
            type: string
      responses:
        '201':
          description: ''
      tags:
        - Favorites
    delete:
      operationId: FavoritesController_deleteArtistFromFavorites
      parameters:
        - name: id
          required: true
          in: path
          example: 68280d15-5093-4e0e-80fd-147e796e252c
          schema:
            type: string
      responses:
        '204':
          description: ''
      tags:
        - Favorites
info:
  title: Library service
  description: The library API description
  version: '1.0'
  contact: {}
tags: []
servers: []
components:
  schemas:
    CreateUserDto:
      type: object
      properties:
        login:
          type: string
          example: Kaia.Toy77
        password:
          type: string
          example: password
      required:
        - login
        - password
    UpdatePasswordDto:
      type: object
      properties:
        oldPassword:
          type: string
          example: password
        newPassword:
          type: string
          example: mZ7BrIDw4HbOoXd
      required:
        - oldPassword
        - newPassword
    CreateArtistDto:
      type: object
      properties:
        name:
          type: string
          example: Mr. Margie Champlin III
        grammy:
          type: boolean
          example: true
      required:
        - name
        - grammy
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
          example: Nora Carroll
        grammy:
          type: boolean
          example: false
      required:
        - name
        - grammy
    CreateAlbumDto:
      type: object
      properties:
        artistId:
          type: string
          example: e2eb1202-1130-40b4-93c0-6bc216b1fd44
        name:
          type: string
          example: Indian Reservation (The Lament Of The Cherokee Reservation Indian)
        year:
          type: number
          example: 2022
      required:
        - artistId
        - name
        - year
    UpdateAlbumDto:
      type: object
      properties:
        artistId:
          type: string
          example: 24d00f4e-d2fd-4ff5-803e-b5fd023f5d3b
        name:
          type: string
          example: Moonlight Serenade
        year:
          type: number
          example: 2022
      required:
        - name
        - year
    CreateTrackDto:
      type: object
      properties:
        name:
          type: string
          example: Purple Haze
        artistId:
          type: string
          example: edf9471e-b8bf-4584-b2da-70ce59601b76
        albumId:
          type: string
          example: 2bd799f0-990f-4e45-99f6-d167e0a32cb6
        duration:
          type: number
          example: 1522
      required:
        - name
        - artistId
        - albumId
        - duration
    UpdateTrackDto:
      type: object
      properties:
        name:
          type: string
          example: Celebration
        artistId:
          type: string
          example: 95b1696c-7a73-487d-939d-7143aa69ad20
        albumId:
          type: string
          example: 3ab7aceb-07ca-4082-ad48-459bdd61714c
        duration:
          type: number
          example: 1797
      required:
        - name
        - artistId
        - albumId
        - duration